pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snowflake-ml-accelerator"
version = "0.1.0"
description = "Rapid ML development and scaling on Snowflake - Remote dev + ML Jobs in one package"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Snowflake", email = "support@snowflake.com"}
]
keywords = ["snowflake", "ml", "gpu", "remote-dev", "ml-jobs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "snowflake-snowpark-python>=1.17.0",
    "snowflake-ml-python>=1.29.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "cryptography>=41.0.0",
    "snowflake-cli @ git+https://github.com/snowflakedb/snowflake-cli@remote_dev_prpr",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
gpu = [
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
]

[project.urls]
Homepage = "https://github.com/snowflakedb/snowflake-ml-accelerator"
Documentation = "https://github.com/snowflakedb/snowflake-ml-accelerator#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/sfml"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
